fix: add least file to keep feature

+++ scripts/python/删除文件夹下过期文件.py
@@ -2,7 +2,7 @@ import os
 import sys
 import time
 
-def delDir(dir,t=120):
+def delDir(dir,t=120, counter=0):
     #获取文件夹下所有文件和文件夹
     files = os.listdir(dir)
     for file in files:
@@ -11,18 +11,20 @@ def delDir(dir,t=120):
         if os.path.isfile(filePath):
             #最后一次修改的时间
             last = int(os.stat(filePath).st_mtime)
+            #文件计数加1
+            counter += 1
             #上一次访问的时间
             #last = int(os.stat(filePath).st_atime)
             #当前时间
             now = int(time.time())
             print("%d, %d", now, last)
             #删除过期文件
-            if (now - last >= t):
+            if (now - last >= t and counter > 100):
                 os.remove(filePath)
                 print(filePath + " was removed!")
         elif os.path.isdir(filePath):
             #如果是文件夹，继续遍历删除
-            delDir(filePath,t)
+            delDir(filePath,t, 0)
             #如果是空文件夹，删除空文件夹
             if not os.listdir(filePath):
                 os.rmdir(filePath)
@@ -36,5 +38,5 @@ if __name__ == '__main__':
     #获取定期清理时间
     ts = int(sys.argv[3])
     while True:
-        delDir(path,t)
+        delDir(path,t,0)
         time.sleep(ts)
